Rename QueryWithAI components to ChatWithPage and CopyContents
