+++ src/components/CopyContents/index.tsx
